+++ views/default/resources/search/index.php
@@ -1,11 +1,6 @@
 <?php
 
-$search_type = get_input('search_type');
-if (!$search_type) {
-	$segments = _elgg_services()->request->getUrlSegments();
-	$page = array_shift($segments); // 'search'
-	$search_type = array_shift($segments);
-}
+$search_type = elgg_extract('search_type', $vars, get_input('search_type', 'object'));
 
 if (!elgg_view_exists("lists/search/$search_type")) {
 	$search_type = 'object';
@@ -15,16 +10,15 @@ $query = get_input('query');
 if (!isset($query)) {
 	$query = get_input('q', get_input('tag', ''));
 }
-$query = stripslashes($query);
 
-$params = array(
+$params = [
 	'query' => $query,
 	'filter_context' => $search_type,
-);
+];
 
 $title = elgg_echo('search');
 
-elgg_push_breadcrumb($title, 'search');
+elgg_push_breadcrumb($title, elgg_generate_url('default:search'));
 
 $content = elgg_view("lists/search/$search_type", $params);
 if (elgg_is_xhr()) {
@@ -33,14 +27,14 @@ if (elgg_is_xhr()) {
 	$filter = elgg_view('filters/search', $params);
 	$sidebar = elgg_view('sidebars/search', $params);
 
-	$params = array(
+	$layout_params = [
 		'content' => $content,
 		'title' => $title,
 		'filter' => $filter,
-		'sidebar' => $sidebar
-	);
+		'sidebar' => $sidebar,
+	];
 
-	$body = elgg_view_layout('content', $params);
+	$body = elgg_view_layout('default', $layout_params);
 
 	echo elgg_view_page($title, $body);
 }