+++ views/default/search/entity.php
@@ -10,55 +10,69 @@ $owner = $entity->getOwnerEntity();
 $container = $entity->getContainerEntity();
 
 if (!$entity->getVolatileData('search_matched_title')) {
-	$title = search_get_highlighted_relevant_substrings($entity->getDisplayName(), $query);
+	$title = htmlspecialchars($entity->getDisplayName(), ENT_QUOTES, 'UTF-8');
+	if ($query) {
+		$title = preg_replace('/(' . preg_quote($query, '/') . ')/i', '<strong>$1</strong>', $title);
+	}
+
 	$entity->setVolatileData('search_matched_title', $title);
 }
 
 if (!$entity->getVolatileData('search_matched_description')) {
-	$desc = search_get_highlighted_relevant_substrings($entity->description, $query);
+	$desc = elgg_get_excerpt((string) $entity->description, 200);
+	if ($query) {
+		$desc = preg_replace('/(' . preg_quote($query, '/') . ')/i', '<strong>$1</strong>', $desc);
+	}
+
 	$entity->setVolatileData('search_matched_description', $desc);
 }
 
 if (!$entity->getVolatileData('search_matched_extra')) {
+	$fields = [];
+	$prefix = '';
+	$exclude = [];
 
 	switch ($type) {
-		case 'user' :
-			$fields = array_keys(elgg_get_config('profile_fields'));
+		case 'user':
+			$fields = array_keys((array) elgg()->fields->get('user', 'user'));
 			$prefix = 'profile';
-			$exclude = array('name', 'description', 'briefdescription');
+			$exclude = ['name', 'description', 'briefdescription'];
 			break;
-		case 'group' :
-			$fields = array_keys(elgg_get_config('group'));
+		case 'group':
+			$fields = array_keys((array) elgg()->fields->get('group', 'group'));
 			$prefix = 'group';
-			$exclude = array('name', 'description', 'briefdescription');
+			$exclude = ['name', 'description', 'briefdescription'];
 			break;
-		case 'object' :
-			$fields = elgg_get_registered_tag_metadata_names();
+		case 'object':
+			$fields = (array) elgg_get_config('registered_tag_metadata_names');
 			$prefix = 'tag_names';
-			$exclude = array('title', 'description');
+			$exclude = ['title', 'description'];
 			break;
 	}
 
-	$matches = array();
+	$matches = [];
 	foreach ($fields as $field) {
 		if (in_array($field, $exclude)) {
 			continue;
 		}
+
 		$metadata = $entity->$field;
 		if (is_array($metadata)) {
 			foreach ($metadata as $text) {
 				if (stristr($text, $query)) {
-					$matches["$prefix:$field"][] = search_get_highlighted_relevant_substrings($text, $query);
+					$highlighted = preg_replace('/(' . preg_quote($query, '/') . ')/i', '<strong>$1</strong>', htmlspecialchars($text, ENT_QUOTES, 'UTF-8'));
+					$matches["$prefix:$field"][] = $highlighted;
 				}
 			}
 		} else {
 			if (stristr($metadata, $query)) {
-				$matches["$prefix:$field"][] = search_get_highlighted_relevant_substrings($metadata, $query);
+				$highlighted = preg_replace('/(' . preg_quote($query, '/') . ')/i', '<strong>$1</strong>', htmlspecialchars($metadata, ENT_QUOTES, 'UTF-8'));
+				$matches["$prefix:$field"][] = $highlighted;
 			}
 		}
 	}
 
-	$extra = array();
+	$extra = [];
 	foreach ($matches as $label => $match) {
 		$extra[] = elgg_format_element('span', ['class' => 'search-match-extra-label'], elgg_echo($label)) . implode(', ', $match);
 	}
@@ -66,12 +80,13 @@ if (!$entity->getVolatileData('search_matched_extra')) {
 	$entity->setVolatileData('search_matched_extra', implode('<br />', $extra));
 }
 
-$view_order = array();
+$view_order = [];
 if ($subtype) {
 	$view_order[] = "search/$type/$subtype/entity";
 }
+
 $view_order[] = "search/$type/entity";
-$view_order[] = "search/entities/entity";
+$view_order[] = 'search/entities/entity';
 
 foreach ($view_order as $view) {
 	if (elgg_view_exists($view)) {
@@ -87,11 +102,10 @@ if (!$icon) {
 	} elseif ($owner instanceof ElggUser) {
 		$icon = elgg_view_entity_icon($owner, $size);
 	} else if ($container instanceof ElggUser) {
-		// display a generic icon if no owner, though there will probably be
-		// other problems if the owner can't be found.
 		$icon = elgg_view_entity_icon($entity, $size);
 	}
 }
+$icon = $icon ?: '';
 
 $title = $entity->getVolatileData('search_matched_title');
 $description = $entity->getVolatileData('search_matched_description');
@@ -102,21 +116,21 @@ if (!$url) {
 	$url = $entity->getURL();
 }
 
-$title = elgg_view('output/url', array(
+$title = elgg_view('output/url', [
 	'text' => $title,
 	'href' => $url,
 	'class' => 'search-matched-title',
-		));
+]);
 
-$subtitle = array();
+$subtitle = [];
 
 if ($subtype) {
-	$type_keys = array(
+	$type_keys = [
 		"$type:$subtype",
 		"$type:default",
 		"item:$type:$subtype",
 		"item:$type",
-	);
+	];
 	foreach ($type_keys as $key) {
 		if (elgg_language_key_exists($key)) {
 			$subtitle['type'] = elgg_echo($key);
@@ -125,20 +139,21 @@ if ($subtype) {
 	}
 }
 
-$byline = array();
+$byline = [];
 if ($type == 'object') {
 	if ($owner) {
-		$owner_link = elgg_view('output/url', array(
+		$owner_link = elgg_view('output/url', [
 			'text' => $owner->getDisplayName(),
 			'href' => $owner->getURL(),
-		));
+		]);
 		$byline[] = elgg_echo('search:owner', [$owner_link]);
 	}
+
 	if ($container && !$container instanceof ElggUser && $container->guid != $owner->guid) {
-		$container_link = elgg_view('output/url', array(
+		$container_link = elgg_view('output/url', [
 			'text' => $container->getDisplayName(),
 			'href' => $container->getURL(),
-		));
+		]);
 		$byline[] = elgg_echo('search:container', [$container_link]);
 	}
 }
@@ -148,6 +163,7 @@ if ($type == 'object') {
 	if (!$time) {
 		$time = elgg_view_friendly_time($entity->time_created);
 	}
+
 	$byline[] = $time;
 }
 
@@ -155,12 +171,12 @@ if (!empty($byline)) {
 	$subtitle['byline'] = implode(' ', $byline);
 }
 
-$last_action = $entity->getVolatileData('select:last_action') ? : max($entity->last_action, $entity->last_login, $entity->time_created);
+$last_action = $entity->getVolatileData('select:last_action') ?: max($entity->last_action, $entity->time_created);
 if ($last_action) {
-	$subtitle['last_action'] = elgg_echo("search:last_action", [elgg_get_friendly_time($last_action)]);
+	$subtitle['last_action'] = elgg_echo('search:last_action', [elgg_get_friendly_time($last_action)]);
 }
 
-$subtitle = elgg_trigger_plugin_hook('subtitle', "search:$type:$subtype", $vars, $subtitle);
+$subtitle = elgg_trigger_event_results('subtitle', "search:$type:$subtype", $vars, $subtitle);
 
 $subtitle_str = '';
 foreach ($subtitle as $s) {
@@ -171,15 +187,17 @@ $content = '';
 if ($description) {
 	$content .= elgg_format_element('div', ['class' => 'search-matched-description'], $description);
 }
+
 if ($extra_info) {
 	$content .= elgg_format_element('div', ['class' => 'search-matched-extra'], $extra_info);
 }
-$summary = elgg_view("$type/elements/summary", array(
+
+$summary = elgg_view("$type/elements/summary", [
 	'entity' => $entity,
 	'tags' => false,
 	'title' => $title,
 	'subtitle' => $subtitle_str,
 	'content' => $content,
-		));
+]);
 
 echo elgg_view_image_block($icon, $summary);